[cleanup] Refactor integration-tests directory
