+++ integration-tests/src/lib.rs
@@ -1,8 +1,8 @@
 pub mod node;
-pub mod runtime_utils;
 pub mod user;
 
 #[cfg(test)]
 mod test_loop;
 #[cfg(test)]
 mod tests;
+pub mod utils;
