+++ integration-tests/src/tests/features/chunk_nodes_cache.rs
@@ -1,4 +1,3 @@
-use crate::tests::client::process_blocks::{deploy_test_contract, set_block_protocol_version};
 use assert_matches::assert_matches;
 use near_chain::Provenance;
 use near_chain_configs::Genesis;
@@ -16,6 +15,8 @@ use near_primitives::version::{ProtocolFeature, ProtocolVersion};
 use near_primitives::views::FinalExecutionStatus;
 use near_vm_runner::logic::TrieNodesCount;
 use nearcore::test_utils::TestEnvNightshadeSetupExt;
+
+use crate::utils::process_blocks::{deploy_test_contract, set_block_protocol_version};
 
 fn process_transaction(
     env: &mut TestEnv,
