+++ integration-tests/src/tests/features/restrict_tla.rs
@@ -1,4 +1,3 @@
-use super::super::process_blocks::create_account;
 use near_chain_configs::Genesis;
 use near_client::test_utils::TestEnv;
 use near_primitives::errors::{ActionError, ActionErrorKind};
@@ -6,6 +5,8 @@ use near_primitives::types::{AccountId, BlockHeight};
 use near_primitives::views::FinalExecutionStatus;
 use near_primitives_core::version::PROTOCOL_VERSION;
 use nearcore::test_utils::TestEnvNightshadeSetupExt;
+
+use crate::utils::process_blocks::create_account;
 
 #[test]
 fn test_create_top_level_accounts() {
