+++ integration-tests/src/tests/features/zero_balance_account.rs
