+++ integration-tests/src/tests/mod.rs
@@ -1,13 +1,7 @@
 mod client;
-mod dependencies;
-mod genesis_helpers;
+mod features;
 mod nearcore;
 mod network;
+mod other_tests;
 mod runtime;
 mod standard_cases;
-mod test_catchup;
-mod test_errors;
-mod test_helpers;
-mod test_overflows;
-mod test_simple;
-mod test_tps_regression;
