+++ integration-tests/src/tests/other_tests/mod.rs
@@ -1,0 +1,6 @@
+mod dependencies;
+mod test_catchup;
+mod test_errors;
+mod test_overflows;
+mod test_simple;
+mod test_tps_regression;
